changed code to multiple file format
