+++ main.py
@@ -1,0 +1,3 @@
+import menu as men
+
+men.home_screen()
